Added a dataAccess module and moved all relevant functions from database.js
