+++ scripts/KneelDiamonds.js
@@ -3,7 +3,7 @@ import { JewelryStyles } from "./JewelryStyles.js";
 import { Orders } from "./Orders.js";
 import { Metals } from "./Metals.js";
 import { Types } from "./Types.js";
-import { addCustomOrder } from "./database.js";
+import { addCustomOrder } from "./dataAccess.js";
 
 document.addEventListener("click", (event) => {
     if (event.target.id === "orderButton") {
